fixing reaction routes in gene page
